refactor(chat): modularize chat functionality by introducing custom hooks for auto-scrolling, chat composition, and session management
