+++ components/chat/chat-page.tsx
@@ -1,12 +1,9 @@
 'use client';
 
-import { useCallback, useEffect, useLayoutEffect, useRef, useState } from 'react';
-import { useRouter } from 'next/navigation';
 import { ArrowDown, Ellipsis, Trash2 } from 'lucide-react';
 import ChatComposer from './chat-composer';
 import Messages from './messages';
-import { useChat } from '@ai-sdk/react';
-import { DbMessage } from '@/lib/ai/types';
+import useChatPage from './use-chat-page';
 import AppHeader from '@/components/layout/app-header';
 import { cn } from '@/lib/utils';
 
@@ -15,246 +12,30 @@ type Props = {
 };
 
 export default function ChatPage({ chatId }: Props) {
-  const router = useRouter();
-  const { messages, setMessages, sendMessage, status, stop } = useChat({
-    onFinish: () => {
-      window.dispatchEvent(new Event('chats:refresh'));
-    },
-  });
-  const [input, setInput] = useState('');
-  const [isActionsOpen, setIsActionsOpen] = useState(false);
-  const actionsMenuRef = useRef<HTMLDivElement | null>(null);
-  const scrollContainerRef = useRef<HTMLDivElement | null>(null);
-  const messagesEndRef = useRef<HTMLDivElement | null>(null);
-  const composerRef = useRef<HTMLDivElement | null>(null);
-  const [composerHeight, setComposerHeight] = useState(56);
-  const [isAtBottom, setIsAtBottom] = useState(true);
-  const [streamReserveHeight, setStreamReserveHeight] = useState(0);
-  const isGenerating = status === 'submitted' || status === 'streaming';
-  const composerOverlayHeight = composerHeight + 10;
-  const messagesBottomPadding = composerHeight + 100;
-  const composerBottomOffset = 32;
-  const disclaimerLineHeight = 16;
-  const disclaimerBottomOffset = composerBottomOffset / 2 - disclaimerLineHeight / 2;
-  const showScrollControl = isGenerating || !isAtBottom;
-
-  const isAtMessagesBottom = useCallback(() => {
-    const scrollContainer = scrollContainerRef.current;
-    const messagesEnd = messagesEndRef.current;
-
-    if (!scrollContainer || !messagesEnd) return true;
-
-    const containerRect = scrollContainer.getBoundingClientRect();
-    const messagesEndRect = messagesEnd.getBoundingClientRect();
-    const visibleBottom = containerRect.bottom - composerOverlayHeight;
-
-    return messagesEndRect.top <= visibleBottom + 8;
-  }, [composerOverlayHeight]);
-
-  const updateBottomState = useCallback(() => {
-    setIsAtBottom(isAtMessagesBottom());
-  }, [isAtMessagesBottom]);
-
-  const scrollToMessagesBottom = useCallback(
-    ({
-      reserveAssistantSpace = false,
-      behavior = 'smooth',
-    }: {
-      reserveAssistantSpace?: boolean;
-      behavior?: ScrollBehavior;
-    } = {}) => {
-      const scrollContainer = scrollContainerRef.current;
-
-      if (!scrollContainer) return;
-
-      if (reserveAssistantSpace) {
-        setStreamReserveHeight(
-          Math.max(140, scrollContainer.clientHeight - messagesBottomPadding - 150),
-        );
-      } else {
-        setStreamReserveHeight(0);
-      }
-
-      window.requestAnimationFrame(() => {
-        const latestScrollContainer = scrollContainerRef.current;
-
-        if (!latestScrollContainer) return;
-
-        latestScrollContainer.scrollTo({
-          top: latestScrollContainer.scrollHeight,
-          behavior,
-        });
-        window.requestAnimationFrame(updateBottomState);
-      });
-    },
-    [messagesBottomPadding, updateBottomState],
-  );
-
-  const scrollToMessagesBottomRef = useRef(scrollToMessagesBottom);
-
-  useLayoutEffect(() => {
-    scrollToMessagesBottomRef.current = scrollToMessagesBottom;
-  }, [scrollToMessagesBottom]);
-
-  useEffect(() => {
-    const handlePointerDown = (event: MouseEvent) => {
-      if (!actionsMenuRef.current?.contains(event.target as Node)) {
-        setIsActionsOpen(false);
-      }
-    };
-
-    document.addEventListener('mousedown', handlePointerDown);
-
-    return () => {
-      document.removeEventListener('mousedown', handlePointerDown);
-    };
-  }, []);
-
-  useEffect(() => {
-    const composer = composerRef.current;
-
-    if (!composer) return;
-
-    const observer = new ResizeObserver(([entry]) => {
-      setComposerHeight(Math.ceil(entry.contentRect.height));
-    });
-
-    observer.observe(composer);
-
-    return () => {
-      observer.disconnect();
-    };
-  }, []);
-
-  useEffect(() => {
-    const scrollContainer = scrollContainerRef.current;
-
-    if (!scrollContainer) return;
-
-    let frame = 0;
-    const handleScroll = () => {
-      window.cancelAnimationFrame(frame);
-      frame = window.requestAnimationFrame(updateBottomState);
-    };
-
-    updateBottomState();
-    scrollContainer.addEventListener('scroll', handleScroll, { passive: true });
-    window.addEventListener('resize', handleScroll);
-
-    return () => {
-      window.cancelAnimationFrame(frame);
-      scrollContainer.removeEventListener('scroll', handleScroll);
-      window.removeEventListener('resize', handleScroll);
-    };
-  }, [updateBottomState]);
-
-  useEffect(() => {
-    const frame = window.requestAnimationFrame(updateBottomState);
-
-    return () => {
-      window.cancelAnimationFrame(frame);
-    };
-  }, [messages, composerHeight, streamReserveHeight, updateBottomState]);
-
-  useEffect(() => {
-    if (isGenerating) return;
-
-    const frame = window.requestAnimationFrame(() => {
-      setStreamReserveHeight(0);
-    });
-
-    return () => {
-      window.cancelAnimationFrame(frame);
-    };
-  }, [isGenerating]);
-
-  useEffect(() => {
-    let cancelled = false;
-    async function loadMessages() {
-      try {
-        const res = await fetch(`/api/messages/${chatId}`);
-
-        if (!res.ok) {
-          console.error('Load messages failed');
-          return;
-        }
-
-        const data = await res.json();
-
-        if (cancelled) return;
-
-        const uiMessages = data.map((message: DbMessage) => ({
-          id: message.id,
-          role: message.role as 'user' | 'assistant' | 'system',
-          parts: [
-            {
-              type: 'text',
-              text: message.content,
-            },
-          ],
-        }));
-
-        setMessages(uiMessages);
-
-        if (cancelled) return;
-
-        const pendingMessageKey = `chat:${chatId}:pending-message`;
-        const pendingMessage = sessionStorage.getItem(pendingMessageKey);
-
-        if (!pendingMessage) return;
-
-        sessionStorage.removeItem(pendingMessageKey);
-        sendMessage({ text: pendingMessage }, { body: { chatId } });
-        scrollToMessagesBottomRef.current({ reserveAssistantSpace: true, behavior: 'smooth' });
-      } catch (error) {
-        console.error('Load messages error:', error);
-      }
-    }
-
-    loadMessages();
-
-    return () => {
-      cancelled = true;
-    };
-  }, [chatId, setMessages, sendMessage]);
-
-  const triggerSend = () => {
-    const text = input.trim();
-
-    if (text === '') return;
-
-    sendMessage(
-      { text },
-      {
-        body: {
-          chatId,
-        },
-      },
-    );
-
-    setInput('');
-    scrollToMessagesBottom({ reserveAssistantSpace: true, behavior: 'smooth' });
-  };
-
-  const deleteChat = async () => {
-    const confirmed = window.confirm('Delete this chat?');
-
-    if (!confirmed) return;
-
-    const res = await fetch(`/api/chats/${chatId}`, {
-      method: 'DELETE',
-    });
-
-    if (!res.ok) {
-      console.error('Delete chat failed');
-      return;
-    }
-
-    setIsActionsOpen(false);
-    window.dispatchEvent(new Event('chats:refresh'));
-    router.push('/');
-    router.refresh();
-  };
+  const {
+    actionsMenuRef,
+    isActionsOpen,
+    setIsActionsOpen,
+    scrollContainerRef,
+    messagesEndRef,
+    composerRef,
+    messages,
+    status,
+    stop,
+    input,
+    setInput,
+    triggerSend,
+    deleteChat,
+    composerHeight,
+    composerOverlayHeight,
+    messagesBottomPadding,
+    streamReserveHeight,
+    showScrollControl,
+    disclaimerBottomOffset,
+    composerBottomOffset,
+    scrollToMessagesBottom,
+    isGenerating,
+  } = useChatPage({ chatId });
 
   return (
     <div
@@ -329,8 +110,13 @@ export default function ChatPage({ chatId }: Props) {
               type="button"
               aria-label={isGenerating ? 'Scroll to latest response' : 'Scroll to bottom'}
               className={cn(
-                'pointer-events-auto flex h-8 items-center justify-center border border-black/8 bg-white/35 text-foreground backdrop-blur-md transition-colors hover:bg-white/50 dark:border-white/10 dark:bg-[#242424]/28 dark:text-white dark:hover:bg-[#242424]/40',
-                isGenerating ? 'w-12 rounded-full' : 'w-10 h-10 rounded-full',
+                'pointer-events-auto relative isolate overflow-hidden',
+                'flex items-center justify-center rounded-full',
+                'border border-black/10 text-black shadow-[0_1px_2px_rgba(0,0,0,0.12),0_3px_10px_rgba(0,0,0,0.10)]',
+                'bg-white/35 backdrop-blur-[2px] backdrop-saturate-150',
+                'transition-colors hover:bg-white/45',
+                'dark:border-white/10 dark:bg-[#2f2f2f]/35 dark:text-white dark:hover:bg-[#2f2f2f]/45',
+                isGenerating ? 'h-9 w-12' : 'h-10 w-10',
               )}
               onClick={() =>
                 scrollToMessagesBottom({
